chore(DTFS2-7727): refactor clone deal tests
