+++ e2e-tests/gef/cypress/e2e/partials/index.js
@@ -8,3 +8,8 @@ export const headingCaption = () => cy.get('[data-cy="heading-caption"]');
 export const mainHeading = () => cy.get('[data-cy="main-heading"]');
 export const saveAndReturnButton = () => cy.get('[data-cy="save-and-return-button"]');
 export const submitButton = () => cy.get('[data-cy="submit-button"]', { timeout: 60000 });
+export const dashboard = () => cy.get('[data-cy="dashboard"]');
+export const dashboardSubNavigation = {
+  deals: () => cy.get('[data-cy="dashboard-sub-nav-link-deals"]'),
+  facilities: () => cy.get('[data-cy="dashboard-sub-nav-link-facilities"]'),
+};
